* master: (43 commits)
  Update release date for v4.2.2
  Fix printf format type for cvode
  Update changelog with release candidate PRs
  Replace repeated bounds-checking with TRACE macro
  Avoid calls to virtual functions in constructors/destructors
  Remove useless cast on nullptr
  Remove dead store in Coordinates, narrow scope of zperiod
  Remove unused private members from BoutMask
  Check after monitor-calls again for user_requested_exit
  Prevent data corruption
  Check if we need to link against libdl for dladdr
  Fix a couple of small(ish) bugs in FV::Div_par_K_Grad_par
  Update changelog
  Bump Zenodo DOI (reserved, unpublished)
  Bump version to 4.2.2
  Fix FV::Div_a_Laplace_perp and FV::Div_par_K_Grad_par
  Initialise some mesh variables to sensible defaults
  Marking some unused args in LaplaceXY
  Fix bug in LaplaceXZ constructor
  Fix prototype Div(v2d, f2d) in header
  ...

+++ src/field/fieldgenerators.hxx
@@ -74,8 +74,10 @@ class FieldGenOneArg : public FieldGenerator { ///< Template for single-argument
 public:
   FieldGenOneArg(FieldGeneratorPtr g) : gen(g) {}
   FieldGeneratorPtr clone(const std::list<FieldGeneratorPtr> args) override {
-    if(args.size() != 1) {
-      throw ParseException("Incorrect number of arguments to function. Expecting 1, got %d", args.size());
+    if (args.size() != 1) {
+      throw ParseException(
+          "Incorrect number of arguments to function. Expecting 1, got %lu",
+          static_cast<unsigned long>(args.size()));
     }
     return std::make_shared<FieldGenOneArg<Op>>(args.front());
   }
@@ -99,7 +101,8 @@ public:
   FieldGeneratorPtr clone(const std::list<FieldGeneratorPtr> args) override {
     if (args.size() != 2) {
       throw ParseException(
-          "Incorrect number of arguments to function. Expecting 2, got %d", args.size());
+          "Incorrect number of arguments to function. Expecting 2, got %lu",
+          static_cast<unsigned long>(args.size()));
     }
     return std::make_shared<FieldGenTwoArg<Op>>(args.front(), args.back());
   }
@@ -125,8 +128,8 @@ public:
       return std::make_shared<FieldATan>(args.front(), args.back());
     }
     throw ParseException(
-        "Incorrect number of arguments to atan function. Expecting 1 or 2, got %d",
-        args.size());
+        "Incorrect number of arguments to atan function. Expecting 1 or 2, got %lu",
+        static_cast<unsigned long>(args.size()));
   }
   BoutReal generate(double x, double y, double z, double t) override {
     if (B == nullptr)
