+++ src/sys/expressionparser.cxx
@@ -86,7 +86,8 @@ public:
 
 FieldGeneratorPtr FieldBinary::clone(const list<FieldGeneratorPtr> args) {
   if (args.size() != 2)
-    throw ParseException("Binary operator expecting 2 arguments. Got '%d'", args.size());
+    throw ParseException("Binary operator expecting 2 arguments. Got '%lu'",
+                         static_cast<unsigned long>(args.size()));
 
   return std::make_shared<FieldBinary>(args.front(), args.back(), op);
 }
